Pather.js is a module now
