+++ d2bs/kolbot/D2BotLead.dbj
@@ -110,7 +110,6 @@ include("automule.js");
 include("gambling.js");
 include("craftingsystem.js");
 include("torchsystem.js");
-include("common/misc.js");
 
 const HeartBeat = require('HeartBeat');
 if (!FileTools.exists("data/" + me.profile + ".json")) {
