+++ d2bs/kolbot/D2BotMule.dbj
@@ -438,7 +438,6 @@ include("automule.js");
 include("mulelogger.js");
 include("torchsystem.js");
 include("common/pather.js");
-include("common/misc.js");
 include("common/prototypes.js");
