+++ d2bs/kolbot/D2BotMuleLog.dbj
@@ -61,7 +61,6 @@ include('require.js');
 include("OOG.js");
 include("MuleLogger.js");
 include("gambling.js");
-include("common/misc.js");
 
 if (!FileTools.exists("data/" + me.profile + ".json")) {
 	DataFile.create();
