+++ d2bs/kolbot/D2BotPublicJoin.dbl
@@ -45,7 +45,6 @@ print('ÿc2Jaensterÿc0 :: Started D2BotPublicJoin.dbj');
 include('require.js');
 include("sdk.js");
 include("OOG.js");
-include("common/misc.js");
 let handle, gameInfo;
 
 function main() {
