+++ d2bs/kolbot/D2BotTCPIP.dbj
@@ -49,7 +49,6 @@ print('ÿc2Jaensterÿc0 :: Started D2BotTcpJoin.dbj');
 include('require.js');
 include("sdk.js");
 include("OOG.js");
-include("common/misc.js");
 include("common/prototypes.js");
 
 function main() {
