+++ d2bs/kolbot/libs/modules/Loader.js
@@ -59,6 +59,7 @@
 		const Messaging = require('Messaging');
 		const Town = require('Town');
 		const Misc = require('Misc');
+		const Pather = require('Pather');
 
 		if (Loader.fileList.indexOf(script) < 0) {
 			Misc.errorReport("ÿc1Script " + script + " doesn't exist.");
