+++ d2bs/kolbot/libs/modules/Paladin.js
@@ -1,4 +1,5 @@
 (function (module, require) {
+	const Pather = require('Pather');
 	const CollMap = require('CollMap');
 	const Paladin = {};
 	const Misc = require('Misc');
