+++ d2bs/kolbot/libs/unit/Movement.js
@@ -4,6 +4,7 @@
  */
 (function (require) {
 	Unit.prototype.moveTo = function (offX, offY, clearPath, pop) {
+		const Pather = require('Pather');
 		this.distance > 5 && Pather.moveTo(this.x + (offX || 0), this.y + (offY || 0), undefined, clearPath, pop);
 		return this;
 	};
@@ -13,11 +14,13 @@
 	};
 
 	PresetUnit.prototype.moveTo = function (offX, offY, clearPath, pop) {
+		const Pather = require('Pather');
 		Pather.moveTo(this.roomx * 5 + this.x + (offX || 0), this.roomy * 5 + this.y + (offY || 0), 4, clearPath, pop);
 		return this;
 	};
 
 	Unit.prototype.bestSpot = function (distance) {
+		const Pather = require('Pather');
 		const CollMap = require('CollMap');
 		let n, i, coll = 0x04,
 			coords = [],
@@ -46,6 +49,7 @@
 
 	Unit.prototype.getIntoPosition = function (distance, coll, walk) { //ToDo; refactor
 		const CollMap = require('CollMap');
+		const Pather = require('Pather');
 		if (!this.x) {
 			return false;
 		}
