+++ d2bs/kolbot/libs/unit/Object.js
@@ -3,6 +3,7 @@
  * @description Some prototypes on objects
  */
 (function (global) {
+    const Pather = require('Pather');
     let coords = function () {
         if (Array.isArray(this) && this.length > 1) {
             return [this[0], this[1]];
